pkg/profefe: refactoring internals of find-request params parsing
